// src/types/generated/FileSessionSnapshot.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GrantedPath } from "./GrantedPath";

export type FileSessionSnapshot = { mainDirectory: GrantedPath | null, currentDirectory: GrantedPath | null, recentDirectories: Array<GrantedPath>, openFiles: Array<GrantedPath>, activeFile: string | null, };